requirements.txt
pyrogram==2.0.106
tgcrypto
motor
pymongo
fastapi
uvicorn[standard]
python-dotenv
aiohttp
